requirements.txt
openai>=1.0.0
matrix-nio[e2e]>=0.24.0
playwright>=1.40.0
duckduckgo-search>=4.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
Pillow>=10.0.0
pymongo[srv]>=4.0.0
